Move MetricRegistry from `c.h.s.api.metrics` into `c.h.s.api` package.
