+++ components/api/src/main/java/com/hotels/styx/api/MetricRegistry.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.metrics;
+package com.hotels.styx.api;
 
 import com.codahale.metrics.Counter;
 import com.codahale.metrics.Gauge;
