+++ components/client/src/main/java/com/hotels/styx/client/OriginsInventory.java
@@ -30,7 +30,7 @@ import com.hotels.styx.api.extension.Origin;
 import com.hotels.styx.api.extension.OriginsChangeListener;
 import com.hotels.styx.api.extension.OriginsSnapshot;
 import com.hotels.styx.api.extension.RemoteHost;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.api.metrics.codahale.CodaHaleMetricRegistry;
 import com.hotels.styx.api.extension.service.BackendService;
 import com.hotels.styx.client.healthcheck.OriginHealthStatusMonitor;
