+++ components/client/src/main/java/com/hotels/styx/client/connectionpool/ConnectionPools.java
@@ -16,7 +16,7 @@
 package com.hotels.styx.client.connectionpool;
 
 import com.hotels.styx.api.extension.Origin;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.api.extension.service.BackendService;
 import com.hotels.styx.client.netty.connectionpool.NettyConnectionFactory;
