+++ components/client/src/main/java/com/hotels/styx/client/connectionpool/StatsReportingConnectionPool.java
@@ -18,7 +18,7 @@ package com.hotels.styx.client.connectionpool;
 import com.codahale.metrics.Gauge;
 import com.hotels.styx.client.Connection;
 import com.hotels.styx.api.extension.Origin;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.api.extension.service.ConnectionPoolSettings;
 import org.slf4j.Logger;
 import rx.Observable;
