+++ components/client/src/test/unit/java/com/hotels/styx/client/applications/metrics/OriginMetricsTest.java
@@ -20,7 +20,7 @@ import com.codahale.metrics.Counter;
 import com.codahale.metrics.Meter;
 import com.hotels.styx.api.Id;
 import com.hotels.styx.api.extension.Origin;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.api.metrics.codahale.CodaHaleMetricRegistry;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
