+++ components/common/src/test/java/com/hotels/styx/api/metrics/codahale/CodaHaleMetricRegistryTest.java
@@ -21,7 +21,7 @@ import com.codahale.metrics.Histogram;
 import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistryListener;
 import com.codahale.metrics.Timer;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
