+++ components/proxy/src/main/java/com/hotels/styx/admin/dashboard/DashboardData.java
@@ -27,7 +27,7 @@ import com.google.common.eventbus.Subscribe;
 import com.hotels.styx.Version;
 import com.hotels.styx.api.extension.OriginsSnapshot;
 import com.hotels.styx.api.extension.OriginsChangeListener;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.api.extension.service.BackendService;
 import com.hotels.styx.api.extension.service.spi.Registry;
