+++ components/proxy/src/main/java/com/hotels/styx/admin/handlers/JVMMetricsHandler.java
@@ -25,7 +25,7 @@ import com.codahale.metrics.MetricRegistryListener;
 import com.codahale.metrics.Timer;
 import com.codahale.metrics.json.MetricsModule;
 import com.google.common.base.Predicate;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 
 import java.time.Duration;
 import java.util.Optional;
