+++ components/proxy/src/main/java/com/hotels/styx/proxy/ProxyConnectorFactory.java
@@ -18,7 +18,7 @@ package com.hotels.styx.proxy;
 import com.codahale.metrics.Histogram;
 import com.hotels.styx.server.HttpErrorStatusListener;
 import com.hotels.styx.api.HttpHandler;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.server.RequestStatsCollector;
 import com.hotels.styx.proxy.encoders.ConfigurableUnwiseCharsEncoder;
 import com.hotels.styx.server.HttpConnectorConfig;
