+++ components/proxy/src/main/java/com/hotels/styx/proxy/plugin/PluginEnvironment.java
@@ -18,7 +18,7 @@ package com.hotels.styx.proxy.plugin;
 import com.codahale.metrics.health.HealthCheckRegistry;
 import com.hotels.styx.api.Environment;
 import com.hotels.styx.api.configuration.Configuration;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.api.plugins.spi.PluginFactory;
 import com.hotels.styx.spi.config.SpiExtension;
