+++ components/proxy/src/test/java/com/hotels/styx/proxy/HttpErrorStatusMetricsTest.java
@@ -20,7 +20,7 @@ import com.codahale.metrics.MetricFilter;
 import com.hotels.styx.server.HttpErrorStatusListener;
 import com.hotels.styx.api.HttpRequest;
 import com.hotels.styx.api.HttpResponseStatus;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.api.metrics.codahale.CodaHaleMetricRegistry;
 import com.hotels.styx.api.plugins.spi.PluginException;
 import org.testng.annotations.BeforeMethod;
