+++ components/proxy/src/test/java/com/hotels/styx/proxy/plugin/PluginEnvironmentTest.java
@@ -18,7 +18,7 @@ package com.hotels.styx.proxy.plugin;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.IntNode;
 import com.hotels.styx.api.Environment;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.api.metrics.codahale.CodaHaleMetricRegistry;
 import com.hotels.styx.api.plugins.spi.PluginFactory;
 import com.hotels.styx.spi.config.SpiExtension;
