+++ components/server/src/test/java/com/hotels/styx/server/netty/handlers/ChannelStatisticsHandlerTest.java
@@ -15,7 +15,7 @@
  */
 package com.hotels.styx.server.netty.handlers;
 
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.api.metrics.codahale.CodaHaleMetricRegistry;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandlerContext;
