+++ support/test-api/src/main/java/com/hotels/styx/testapi/StyxServer.java
@@ -24,7 +24,7 @@ import com.hotels.styx.api.Environment;
 import com.hotels.styx.api.extension.Origin;
 import com.hotels.styx.api.configuration.Configuration;
 import com.hotels.styx.api.configuration.Configuration.MapBackedConfiguration;
-import com.hotels.styx.api.metrics.MetricRegistry;
+import com.hotels.styx.api.MetricRegistry;
 import com.hotels.styx.api.plugins.spi.Plugin;
 import com.hotels.styx.api.plugins.spi.PluginFactory;
 import com.hotels.styx.infrastructure.MemoryBackedRegistry;
